Extract Audit from model_soup
