+++ grouper/fe/handlers/audits_view.py
@@ -1,7 +1,7 @@
 from grouper.audit import get_audits
 from grouper.constants import AUDIT_MANAGER, AUDIT_VIEWER
 from grouper.fe.util import GrouperHandler
-from grouper.model_soup import Audit
+from grouper.models.audit import Audit
 from grouper.models.audit_log import AuditLog, AuditLogCategory
 from grouper.user_permissions import user_has_permission
