+++ grouper/user.py
@@ -3,8 +3,9 @@ from datetime import datetime
 from sqlalchemy import or_
 from sqlalchemy.sql import label, literal
 
-from grouper.model_soup import (APPROVER_ROLE_INDICIES, Audit, Group, GROUP_EDGE_ROLES, GroupEdge,
+from grouper.model_soup import (APPROVER_ROLE_INDICIES, Group, GROUP_EDGE_ROLES, GroupEdge,
     OWNER_ROLE_INDICES, Request, RequestStatusChange)
+from grouper.models.audit import Audit
 from grouper.models.audit_log import AuditLog
 from grouper.models.base.session import Session  # noqa
 from grouper.models.comment import Comment
